fix: pet configs reading as int instead of double

+++ UltimatePets/src/main/java/mc/ultimatecore/pets/configs/Config.java
@@ -42,8 +42,8 @@ public class Config extends YAMLFile {
         selectedPetNone = getConfig().getString("selectedPetNone");
         for(Tier tier : HyperPets.getInstance().getTiers().getTierList().values())
             petItems.put(tier, Utils.getItemFromConfig(getConfig(), "petItems."+tier.getName()));
-        skillsXP = (int) getConfig().getDouble("xp.skillsPoints");
-        blockBreakXP = (int) getConfig().getDouble("xp.blockBreak");
+        skillsXP = getConfig().getDouble("xp.skillsPoints");
+        blockBreakXP = getConfig().getDouble("xp.blockBreak");
         debug = getConfig().getBoolean("DEBUG");
         syncDelay = getConfig().getInt("database.syncDelay");
         //------------------------------------------------//
